requirements.txt
fastapi
uvicorn
mysql-connector-python
bcrypt
email-validator
